+++ web/src/components/MainNavigator.jsx
@@ -1,26 +1,36 @@
 import { createStackNavigator } from "@react-navigation/stack";
+import { createDrawerNavigator } from "@react-navigation/drawer";
 import React from "react";
 import { ROUTES } from "../constants";
 import Front from "./Front";
 import Login from "./Login";
 
-const Stack = createStackNavigator();
-
 const MainNavigator = () => {
+  const Drawer = createDrawerNavigator();
   return (
     <>
-      <Stack.Navigator initialRouteName={ROUTES.LOGIN}>
-        <Stack.Screen
+      <Drawer.Navigator 
+            initialRouteName={ROUTES.LOGIN}
+             screenOptions={{
+              headerShown: false,
+              drawerActiveBackgroundColor: "#346633",
+              drawerActiveTintColor: "white",
+              drawerLabelStyle: {
+                marginLeft: 0,
+              },
+            }}
+      >
+        <Drawer.Screen
           name={ROUTES.LOGIN}
           component={Login}
           options={{ headerShown: false }}
         />
-        <Stack.Screen
+        <Drawer.Screen
           name={ROUTES.FRONT}
           component={Front}
           options={{ headerShown: false }}
         />
-      </Stack.Navigator>
+      </Drawer.Navigator>
     </>
   );
 };
