+++ web/src/components/Navbar.jsx
@@ -1,7 +1,9 @@
 import { useState} from "react";
-import { StyleSheet, Text, TextInput, View, useWindowDimensions } from "react-native";
+import { StyleSheet, Text, TextInput, View, useWindowDimensions, Button } from "react-native";
+import { useNavigation } from '@react-navigation/native';
 
 const Navbar = () => {
+  const nav = useNavigation();
   const { height, width } = useWindowDimensions()
 
   const style = StyleSheet.create({
@@ -34,7 +36,7 @@ const Navbar = () => {
   }
   return ( 
     <View style={style.header}>
-      <Text style={style.headerText}>Menu</Text>
+      <Button title="Menu" style={style.headerText} onPress={() => nav.toggleDrawer()}/>
       <TextInput
           style={style.searchBar}
           onChangeText={() => realtimeSearchDB()}
